requirements.txt
streamlit
mysql-connector-python
bcrypt
pandas
streamlit-option-menu
pyipp